modularize stuff from stream
